+++ src/witsec-design-blocks-block/component.js
@@ -3,25 +3,32 @@ mbrApp.loadComponents(
 	{"witsec-design-blocks-block":{
 			_group:"witsec",
 
-			_onParamsShow: function(e,$params,$block) {
+			// If the witsecBlockURL is set AND the block is set to be activated, something must've gone wrong activating the block. Let's try again :)
+			_onLoadFromProject: function(b,	a) {
+				if (this._params.witsecBlockURL && this._params.witsecBlockActivate) {
+					setTimeout(function() {
+						mbrApp.alertDlg("A Design Block hasn't been activated properly. Please activate it manually using the gear icon.");
+					}, 1000);
+				}
 			},
 
 			// When you activate this block, Mobirise will restart automatically.
 			_params:{
-				notice:      {type:"separator",title:"<a href='https://witsec.nl/mobirise/gallery/?version=0.4&type=" + (mbrApp.isAMP() ? "amp" : "bootstrap") + "&random=" + Math.floor(Math.random() * 100000000) + "' target='_blank' class='alert alert-light' style='cursor:hand; color:#000; text-decoration:none'>Design Blocks Gallery</a><br /><br />"},
-				templateURL: {type:"text",title:"Template (paste here)",default:""},
-				activate:    {type:"switch",title:"Activate block",default:!1,condition:["templateURL"]}
+				witsecGallery:       {type:"separator",title:"<a href='#' class='btn btn-primary witsec-show-design-gallery'>OPEN GALLERY</a>"},
+				witsecBlockURL:      {type:"text",title:"Block URL (paste here)",default:""},
+				witsecBlockActivate: {type:"switch",title:"Activate block",default:!1,condition:["witsecBlockURL"]}
 			},
 
-			_message: "Use the gear icon to access the Design Blocks Gallery or paste the location of your own template.",
+			// Default message we show the user
+			_message: "Use the gear icon to access the Design Blocks Gallery or paste the location of your own block.",
 
 			_onParamsChange:function($item, param, val){
-				if (param == "activate" && val == true) {
+				if (param == "witsecBlockActivate" && val == true) {
 					// We add a random number to the URL, to avoid any caching
-					var templateURL = this._params.templateURL + "?" + Math.floor(Math.random() * 100000000);
+					var witsecBlockURL = this._params.witsecBlockURL + "?" + Math.floor(Math.random() * 100000000);
 
-					// Try to grab the template
-					var request = $.ajax({ url: templateURL, dataType: "text" });
+					// Try to grab the block
+					var request = $.ajax({ url: witsecBlockURL, dataType: "text" });
 
 					// On failure
 					request.error(function(jqXHR, textStatus, errorThrown) {
@@ -29,7 +36,7 @@ mbrApp.loadComponents(
 							m = "The server is not responding.";
 
 						if (textStatus == "error")
-							m = "An error occured while loading the template. " + errorThrown;
+							m = "An error occured while loading the block. " + errorThrown;
 
 						mbrApp.alertDlg(m);
 						return false;
@@ -46,7 +53,7 @@ mbrApp.loadComponents(
 							json = jQuery.parseJSON(result);
 						}
 						catch(e) {
-							mbrApp.alertDlg("The template you selected does not appear to be valid JSON.");
+							mbrApp.alertDlg("The block you selected does not appear to be valid JSON.");
 							return false;
 						}
 
@@ -54,7 +61,7 @@ mbrApp.loadComponents(
 						var error = "";
 						var warn = "";
 						if (!json.hasOwnProperty('data'))
-							error = "The template does not contain the required field 'data'.";
+							error = "The block does not contain the required field 'data'.";
 
 						// Check if _styles exists (doesn't have to be there)
 						if (!json.data.hasOwnProperty('_styles'))
@@ -62,14 +69,14 @@ mbrApp.loadComponents(
 
 						// Check _customHTML
 						if (!json.data.hasOwnProperty('_customHTML'))
-							error = "The template does not contain the required field '_customHTML'.";
+							error = "The block does not contain the required field '_customHTML'.";
 						else {
 							if (!json.data._customHTML.match(/<mbr-parameters>/i)) {
 								warn += "<li>There don't appear to be any parameters inside the '_customHTML' field. This means you will not be able to customize this block using the 'gear icon'.</li>";
 							}
 						}
 
-						// In case the template is a custom html block created with the Code Editor, try to convert the CSS into JSON
+						// In case the block is a custom html block created with the Code Editor, try to convert the CSS into JSON
 						if (json.data.hasOwnProperty('_customCSS') && json.data._customCSS != "" && json.data._customTemplate) {
 							// _customCSS exists, but _styles is empty, so let's try to convert CSS to JSON here
 							var css = json.data._customCSS;
@@ -108,13 +115,13 @@ mbrApp.loadComponents(
 								json.data._styles = jsoncss;
 							}
 							catch(e) {
-								warn += "<li>There was an issue processing the '_customCSS' field. The template will be imported, but it may have styling issues.</li>";
+								warn += "<li>There was an issue processing the '_customCSS' field. The block will be imported, but it may have styling issues.</li>";
 							}
 						}
 
 						// If there's an error, show it and return
 						if (error) {
-							mbrApp.alertDlg("<b>Invalid template</b><br /><br />" + error);
+							mbrApp.alertDlg("<b>Invalid block</b><br /><br />" + error);
 							return false;
 						}
 
@@ -132,6 +139,9 @@ mbrApp.loadComponents(
 
 						// Let the user know what's going on
 						block._message = "Block is prepared. Mobirise will reload this project in a few seconds... <img src='" + mbrApp.getAddonsDir() + "/witsec-design-blocks/witsec-design-blocks-block/waiting.gif' />";
+
+						// Hide the components params
+						mbrApp.hideComponentParams();
 
 						// Reload the current project. Required in order to be able to fully use this new block
 						setTimeout(function() {